completed api of transaction history and added mappers

+++ finpayai/src/main/java/com/finpay/finpayai/repo/TransactionRepo.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.*;
 import com.finpay.finpayai.entity.*;
 
 @Repository
-public interface TransactionRepo extends JpaRepository<transactionEntity, Long> {
+public interface TransactionRepo extends JpaRepository<transactionEntity, Integer> {
 }
